+++ packages/oruga-next/src/components/dropdown/Dropdown.vue
@@ -1,12 +1,12 @@
 <script setup lang="ts">
 import { computed, nextTick, ref, watch, type PropType, type Ref } from "vue";
 
-import { baseComponentProps } from "@/utils/SharedProps";
+import PositionWrapper from "../utils/PositionWrapper.vue";
+
 import { getOption } from "@/utils/config";
 import { vTrapFocus } from "@/directives/trapFocus";
 import { toCssDimension, isMobileAgent } from "@/utils/helpers";
 import { isClient } from "@/utils/ssr";
-import PositionWrapper from "@/utils/PositionWrapper.vue";
 import {
     unrefElement,
     defineClasses,
@@ -34,8 +34,8 @@ defineOptions({
 });
 
 const props = defineProps({
-    // add global shared props (will not be displayed in the docs)
-    ...baseComponentProps,
+    /** Override existing theme classes completely */
+    override: { type: Boolean, default: undefined },
     /** @model */
     modelValue: {
         type: [String, Number, Boolean, Object, Array],
@@ -100,15 +100,20 @@ const props = defineProps({
     },
     /** Dropdown will be expanded (full-width) */
     expanded: { type: Boolean, default: false },
+    /** HTML element ID of the dropdown menu element */
+    menuId: { type: String, default: null },
+    /** Tabindex of the dropdown menu element */
+    menuTabindex: { type: Number, default: null },
     /** Dropdown menu tag name */
     menuTag: {
         type: [String, Object, Function] as PropType<DynamicComponent>,
-        default: () => getOption("dropdown.menuTag", "div"),
+        default: () => getOption<DynamicComponent>("dropdown.menuTag", "div"),
     },
     /** Dropdown trigger tag name */
     triggerTag: {
         type: [String, Object, Function] as PropType<DynamicComponent>,
-        default: () => getOption("dropdown.triggerTag", "div"),
+        default: () =>
+            getOption<DynamicComponent>("dropdown.triggerTag", "div"),
     },
     /**
      * Dropdown will be triggered by any events
@@ -140,11 +145,13 @@ const props = defineProps({
     /**
      * Role attribute to be passed to the list container for better accessibility.
      * Use menu only in situations where your dropdown is related to a navigation menu.
-     * @values list, menu, dialog
+     * @values list, listbox, menu, dialog
      */
     role: {
         type: String,
         default: getOption("dropdown.ariaRole", "list"),
+        validator: (value: string) =>
+            ["list", "listbox", "menu", "dialog"].indexOf(value) > -1,
     },
     /** Mobile breakpoint as max-width value */
     mobileBreakpoint: {
@@ -161,46 +168,57 @@ const props = defineProps({
         default: () => getOption("dropdown.teleport", false),
     },
     // class props (will not be displayed in the docs)
+    /** Class of the root element */
     rootClass: {
         type: [String, Array, Function] as PropType<ComponentClass>,
         default: undefined,
     },
+    /** Class when the dropdown is teleported */
     teleportClass: {
         type: [String, Array, Function] as PropType<ComponentClass>,
         default: undefined,
     },
+    /** Class of the trigger element */
     triggerClass: {
         type: [String, Array, Function] as PropType<ComponentClass>,
         default: undefined,
     },
+    /** Class of dropdown menu when inline */
     inlineClass: {
         type: [String, Array, Function] as PropType<ComponentClass>,
         default: undefined,
     },
+    /** Class of the overlay when on mobile */
     menuMobileOverlayClass: {
         type: [String, Array, Function] as PropType<ComponentClass>,
         default: undefined,
     },
+    /** Class of the dropdown menu */
     menuClass: {
         type: [String, Array, Function] as PropType<ComponentClass>,
         default: undefined,
     },
+    /** Class of dropdown menu position */
     menuPositionClass: {
         type: [String, Array, Function] as PropType<ComponentClass>,
         default: undefined,
     },
+    /** Class of dropdown menu when active */
     menuActiveClass: {
         type: [String, Array, Function] as PropType<ComponentClass>,
         default: undefined,
     },
+    /** Class of dropdown when on mobile */
     mobileClass: {
         type: [String, Array, Function] as PropType<ComponentClass>,
         default: undefined,
     },
+    /** Class of dropdown when disabled */
     disabledClass: {
         type: [String, Array, Function] as PropType<ComponentClass>,
         default: undefined,
     },
+    /** Class of dropdown when expanded */
     expandedClass: {
         type: [String, Array, Function] as PropType<ComponentClass>,
         default: undefined,
@@ -506,7 +524,6 @@ const menuClasses = defineClasses(
         autoPosition,
         computed(() => !!autoPosition.value),
     ],
-
     [
         "menuActiveClass",
         "o-drop__menu--active",
@@ -534,7 +551,7 @@ defineExpose({ $trigger: triggerRef, $content: contentRef });
             ref="triggerRef"
             :tabindex="disabled ? null : tabindex"
             :class="triggerClasses"
-            aria-haspopup="true"
+            :aria-haspopup="ariaRole === 'list' ? true : ariaRole"
             @click="onClick"
             @contextmenu="onContextMenu"
             @mouseenter="onHover"
@@ -569,12 +586,14 @@ defineExpose({ $trigger: triggerRef, $content: contentRef });
                 <component
                     :is="menuTag"
                     v-show="(!disabled && (isActive || isHovered)) || inline"
+                    :id="menuId"
                     ref="contentRef"
                     v-trap-focus="trapFocus"
+                    :tabindex="menuTabindex"
                     :class="menuClasses"
                     :aria-hidden="!isActive"
-                    :role="role"
-                    :aria-modal="!inline"
+                    :role="ariaRole"
+                    :aria-modal="!inline && trapFocus"
                     :style="menuStyle">
                     <!--
                         @slot Place dropdown items here